refactor: deepen run history module
